+++ resources/views/layouts/guest.blade.php
@@ -10,6 +10,7 @@
         <link rel="preconnect" href="https://fonts.googleapis.com">
         <link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Caveat:wght@400;600;700&family=IBM+Plex+Mono:wght@300;400;600&display=swap" rel="stylesheet">
 
+        <x-analytics />
         @vite(['resources/css/app.css', 'resources/js/app.js'])
 
         @livewireStyles